+++ README.md
@@ -37,12 +37,12 @@ The backend listens on `http://localhost:3000` to store tasks in a local JSON fi
 
 ### Loading sample tasks
 
-A small list of example tasks is available under `backend/test_tasks.json`. These
+A small list of example tasks is available under `backend/data/test_tasks.json`. These
 are **not** loaded automatically. If you want to populate the database with them
 run:
 
 ```bash
-node backend/add_test_tasks.js
+node backend/scripts/add_test_tasks.js
 ```
 
 This script reads the JSON file and appends the tasks to `backend/db.json`.
