+++ backend/scripts/add_test_tasks.js
@@ -6,13 +6,13 @@ const path = require('path');
   const { JSONFile } = await import('lowdb/node');
   const { v4: uuidv4 } = require('uuid');
 
-  const dbFile = path.join(__dirname, 'db.json');
+  const dbFile = path.join(__dirname, '../db.json');
   const adapter = new JSONFile(dbFile);
   const db = new Low(adapter);
   await db.read();
   db.data ||= { tasks: [], users: [] };
 
-  const tasksPath = path.join(__dirname, 'test_tasks.json');
+  const tasksPath = path.join(__dirname, '../data/test_tasks.json');
   const tasks = JSON.parse(fs.readFileSync(tasksPath, 'utf8'));
 
   tasks.forEach(t => {
