+++ frontend/components/NavigationBar.js
