+++ frontend/components/Tile.js
