+++ frontend/pages/DashboardPage.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useState, useMemo } from 'react';
 import { View, Text, FlatList, StyleSheet } from 'react-native';
 import { IconButton } from 'react-native-paper';
-import Tile from './Tile';
-import { EVENT_COLOR } from './colors';
+import Tile from '../components/Tile';
+import { EVENT_COLOR } from '../utils/colors';
 
 /**
  * Simple dashboard shown on login. It lists upcoming events for the
